NOTE: This checkin contains changes from several separate checkins on the inside. All merging was done manually and may have resulted in changes from the Codeplex files being lost!!
Two new classes for aggregating source model information: one for ISourceLocationProvider and another for ILocalScopeProvider. But in order to be able to aggregate those interfaces, they had to be modified so that for each element the aggregator can retrieve the containing method (definition) and then map that to the appropriate provider (the assumption is that each provider is attached to a particular unit and so the aggregator can climb from the method definition up to the containing unit). Modifying the interfaces induced many consequent changes, mostly to add a method definition as a parameter to classes that implement the interfaces so that they can support that query.

Bugfix: Mutable model's generictypeparameterreference and genericmethodparameterreference classes' resolution always picked the first generic parameter of the defining type/method due to a programming mistake. 

Fix compilation of bound expressions where the bound definition is a static field.

Various fixes to help with decompilation and recompilation of typed references.

Bug fix: generate ldelem/stelem/etc instead of ldelem.ref when the element type is a generic (type or method) parameter.  Previously the generic method parameter case was missed. 

Make source emitter aware of conversions, rethrow statements and finalizers. Fix decompilation of nested try statements that start at the same location but do not protect the same try body. Fix code gen for rethrow statements and catch handlers that immediately pop the exception value.

Fix decompile/compile round trip for box instructions.

Added comment in ILReader for when array types use psuedo instuctions and when not.