[Catching up with the past 11 days of checkins to the internal code base. Some of these comments might be for files in the Metadata project.]
Deal with castclass instructions that cast to value types. (Not legal according to ECMA, but seems to occur all the same.)
The type of a shift expression should be the same as the type of the left operand. 
Restrict the situations in which c ? x : y is printed out as c || y to be exactly c ? 1 : y not c ? any int != 0 : y
Big change to the way contracts are accessed from the Metadata model: the new model is to use an IContractAwareHost as the host if one is interested in contracts. The host is responsible for loading any reference assemblies and attaching an IContractExtractor to each unit that it loads. The IContractExtractor is responsible for decompiling method bodies and then using a ContractExtractor that splits method bodies by removing the code that represents the contracts. Because of the new validator/abbreviator support, getting the contracts from one method may involve decompiling and extracting contracts from other methods. So a callback protocol is used: any client interested in the residual method bodies after contract extraction must register to be notified because the extractor does not keep those bodies.

Contract extraction no longer happens implcitly as part of decompiling a method body, but must be explicitly requested via the IContractExtractor/IContractProvider interfaces.
New validator/abbreviator support: note that they are not deleted from the method body they occur in, but do show up in the contracts extracted from those methods.

1) Do not use ldnull to load zero IntPtr or UIntPtr.
2) Do not lose track of the sign of unsigned popped values during Unstacker.
3) Do not emit nop conversions such as I8 to U8.
4) Use brfalse and brtrue when comparing IntPtr or UIntPtr values to zero.
5) Use comparison branches when dealing with !(op1 comparison op2) kind of expressions in branches.

Deal with expressions that take the address of byref parameters.

Start a new basic block after each endfinally instruction. Output fault blocks in C# format, even though it is not legal C# syntax.

Normalize line-endings.

Introduce an explicit operand stack into the CodeModel to deal with controlled mutability pointer values, which can only live on the operand stack.

Roundtrip filtered catch handlers via the CodeModel.

If the method being decompiled has no local variables, set InitLocalVariables to true in the decompiled version, since it may have temporary locals introduced by Unstacker.

Add dup instructions to CodeModel to deal with pointers that may not be stored into locals but get duplicated on the stack.

Uninstantiate or unspecialize the method before looking up its contract.

making checkin suit to use code mutator

Bug fixes: struct that implements ienumerator, generic method parameter in anonymous delegates decompilation, and the capture and substitution of parameter of nested anonymous delegates.

Validation code (legacy preconditions and calls to Validators/Abbreviators) should not be preserved in contracts on any override or interface/abstract class implementation. They are in the code, but should not be in the contracts. So the Contract Extractor now ignores them (except on the base methods that originally declare them).

When doing a typed stdind instruction, insert a type case on the address if it does not already have a suitable type.

Emit correct Add/Subtract opcode for additions/subtractions of pointers and references by looking at the types of the operands.

Fixed the contract extractor so that it properly skips over the "ctor preamble", i.e., the code corresponding to the base/this call in a constructor.

Renamed the namespace Microsoft.Cci.Contracts to Microsoft.Cci.MutableContracts for the MutableObjectModel.
Added functionality to the CodeContractAwareHostEnvironment so that it reloads units that have been modified since the last time they were loaded.