This fix solves an issue that prevents the correctly replacement of the DoWhile loop with the For loop, when the original source code has a foreach loop that iterates over an array.
For example, for a loop like this:

foreach (var elem in array)
 // some statement

the buggy Decompiler generates something like this:

int tempVar = 0;
while (tempVar < array.Length)
{
 var elem = array[tempVar];
 // some statement
 tempVar++;
}

It is ok, but it I think it would be nicer to generate the following code:

for (int tempVar = 0; tempVar < array.Length; tempVar++)
{
 var elem = array[tempVar];
 // some statement
}

Ideally, the Decompiler could generate the original foreach loop, but with this little fix at least it will generate a for and not a while loop. This bug also happens if the original code is directly the while version.