A bug in Tortoise SVN is preventing me from adding the file Unstacker.cs.!!
Fix the Unstacker so that it converts ints to bools when one of the pushes is of a bool and the corresponding one in the other branch of a conditional is a push of an int.
Fix PeToPe and PeToText to deal with command line args correctly and to use the Unstacker when indicated.
Update CodeModelTest to use the Unstacker.