My recent change related to IsFacade check hit an unfortunate problem in ILC: according to Morgan's suggestion I added a bit of logic into the method which looks at actual content of the IL file - namely at certain attributes saying whether the assembly belongs to the framework assembly set. This logic started indirectly triggering WinRT redirection logic which fails in ILHostEnvironment with nullref in initial assembly closures when they ask whether the assemblies are facades.
Based on offline chat with Vitek the most expedient fix seems to be using the already existing "projectToCLRTypes" parameter to bail out early in the Redirect method and thus avoid hitting the null PlatformWinRTTypeDefinitions as the assembly closure host naturally doesn't load the WinMD types. The problem is strangely non-deterministic, it hit just a couple of test cases in post-checkin although in truth it should hit every test. Perhaps it is somehow affected by the exact framework bits copied to the app as tests generally run in custom framework mode.

Thanks

Tomas

[tfs-changeset: 1634725]