Fix Bug 285745: [Customer] Native binary architecture check seems incorrect for some cases (ARM?)
There is a check in the ILC BuildDriver that compares the architecture of native binaries in the ILC input folder to the target architecture of ILC, and this check may incorrectly say some ARM binaries didn't match the ARM target architecture of ILC. This happened because we would only recognize Machine Type 0x1c0 (ARM little endian) as ARM but we would not recognize Machine Type 0x1c4 (ARM Thumb-2 little endian) as ARM. My changes the check to correctly identify both Machine Types as valid for ARM-targeting ILC's.

Note: The extraneous clrcompression.dll check was removed because this does not appear to be an issue anymore.

[tfs-changeset: 1639002]