Fix corrupt CLR headers in assembly produced by IL2IL in some rare cases.
In PeWriter there's logic which determines if the "blob" blob needs 16bit or 32bit indices. This logic depends on the overall size of the blob being written. The underlying problem is that this logic is called multiple times and the result is used in various places. IN one case it's used to compute the size of the .text section in another it's used to compute the size of the metadata section. Unfortunetly between these two calls the blob is aligned, that is it's padded with zeroes. If the original blob size is near the boundary (0x10000) it may happen that the first calculation returns 16bit indices while the second one (after padding) returns 32bit. This ends up in inconsistent sizes of various headers and effectively corrupt PE/IL file.
This change modifies the existing code for the same problem in the "Strings" case to cover even cases where the original value is 0xFFFE or 0xFFFD. The login now follows the code - compute the size with alignment and test it against ECMA spec values. The same mechanism is then used for both blobs and strings.

[tfs-changeset: 1647012]