Check in rest of portable PDB scaffolding in ILC except for the portable PDB support itself
As portable PDB testing (currently underway) turns out to be non-trivial, I propose checking in most of the remaining changes where I just backed out the actual bits dealing with Tomas Matousek's converter change. So technically this change doesn't change anything in ILC or elsewhere, it just rids me of keeping the private changes around for an extended period of time. The addition to dirs.proj makes sure that, once Cci becomes dependent on Tomas' converter, it will start to get deployed along with it to ILC layouts.

Thanks

Tomas

[tfs-changeset: 1651614]