Final check-in of portable PDB support
Addition of portable PDB support to CCI. As of this update, portable PDB support is internal i.e. it's not normally available in the GitHub build of Microsoft.Cci. If some GitHub users want to unofficially experiment with it, all they have to do is pull down the symreader-converter package from myget and slightly modify the project file. We can make the portable PDB support "official" once it's completely finished (in this version we still don't support embedded portable PDB's) and sufficiently tested.

[tfs-changeset: 1653073]