Implements optimized closure computation to avoid including assemblies which are only referenced by pure facades.
The main driver for this change is the new netstandard.dll which is a pure facade with references to most of the framework. With the old code any reference to this facade would pull in most of the framework regardless if it's needed or not.
The optimization is to not include assemblies referenced by facades by default, instead in such case walk the type references through the facade to pick only the implementation assemblies which are actually needed by anything.

In order to implement this, some additional changes have been made:
- The closure computation has now 2 passes. In the first we don't optimize and walk all the references, loading all the assemblies into the host (Since we have custom loading logic). In the second pass we compute the closure again applying the optimizations.
- New helper method was added to CCI which allows the caller to traverse one type forward at a time. The existing CCI APIs sometimes fully resolved the type forward to the final type definition thus effectively skipping any assemblies in between.

This change has these effects on the resulting closure:
- All references are still reported in the per-assembly reference tables, so even pure facades get the full list
- Assemblies which refer to pure facades may get more references as we include the actual implementation assembly for the type behind the facade (and any intermediates in between).
This is really only important for incremental builds, in which case we do need to know the full list of references to correctly compute which assemblies should be rebuilt.

[tfs-changeset: 1655094]