Support embedded portable PDB's in CCI and .NET Native toolchain
Additional CCI changes and slight refactoring of PdbFormatProvider needed for portable PDB support. This also cleaned up the separation between the pre-existing Windows PDB logic and new portable PDB support so that the entire PdbFormatProvider file is only included in build when portable PDB support is enabled. As of this check-in, portable PDB support in CCI is still "unofficial" [it is turned off in GitHub builds of CCI and turning it on would require additional tweaks to the project files].

[tfs-changeset: 1656159]