Fix ILC compilation of Basic code
Jose noticed that UAPAOT testing of new Visual Basic tests started crashing ILC due to a failed range check in MethodBodyEditor. Based on Tomas Matousek's clarification my understanding is that Visual Basic compiler uses different semantics when emitting local scopes into a legacy Windows PDB - namely it emits the "length" of a local scope one less than Roslyn would in an equivalent situation - in other words, the "scope end" as emitted by the VB compiler corresponds to the last byte of the scope whereas, for Roslyn, the end of the scope is the first byte after the end of the scope.

As according to Tomas this behavior differs between Windows PDB and Portable PDB (when producing portable PDB's, the VB compiler honors the same semantics as Roslyn), I believe the clean fix is to internally correct this in the Windows PDB reader and that is what my fix does based on Tomas' advice. When reading method sequence points, whenever the Basic language GUID is identified, a new internal method on PdbFunction is called and it adjusts all the local scopes. With this change, I see Jose's test ILC-compile just fine.

Thanks

Tomas

[tfs-changeset: 1663545]