Update Microsoft.DiaSymReader and Microsoft.DiaSymReader.Converter
This change update the DiaSymReader and DiaSymReader.Converter. With the latest assemblies the internal PdbWriter<T> is replaced with SymUnmanagedWriter. This change ensures that PdbWriterForCci implements that abstract class. Most of the additional functions of SymUnmanagedWriter looks irrelevant to me from CCI's perspective. But I am not an expert in this subject :). I have run couple of tests with CLRTestPortablePdb and CLRTestEmbeddedPdb turned on and it seems like it is working okay.

[tfs-changeset: 1680609]